[capi] Avoid going through Heat for worker list
By default, Magnum UI goes through Heat to get the list of nodes
which is not correct since it's making an assumption that Heat
is always in use.

The fix for this would be to make sure that Magnum has a list of
all the VMs in it's database (or some sort of API call that
returns them all from the driver) but that's quite a big amount
of work to implement for now.

So for now, if stack_id doesn't look like a UUID, we assume it
is deployed using Clsuter API driver for Magnum and look up with
that alternative method instead.

(cherry picked from commit 6f31cc5cacf23398b76392922ee9863d50aa9e7e)